+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/resource/filter/fluid/FluidFilteredResource.java
@@ -43,6 +43,11 @@ public record FluidFilteredResource(FluidResource value, long amount) implements
     }
 
     @Override
+    public Component getDisplayName() {
+        return Platform.INSTANCE.getFluidRenderer().getDisplayName(value);
+    }
+
+    @Override
     public List<Component> getTooltip() {
         return Platform.INSTANCE.getFluidRenderer().getTooltip(value);
     }