+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/resource/filter/item/ItemFilteredResource.java
@@ -51,6 +51,11 @@ public record ItemFilteredResource(ItemResource value, long amount) implements F
     }
 
     @Override
+    public Component getDisplayName() {
+        return value.toItemStack().getHoverName();
+    }
+
+    @Override
     public List<Component> getTooltip() {
         final Minecraft minecraft = Minecraft.getInstance();
         return ClientProxy.getPlayer().map(player -> value.toItemStack().getTooltipLines(