+++ CHANGELOG.md
@@ -10,11 +10,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 -   Constructor
+-   Filter slot hints.
 
 ### Changed
 
 -   The Constructor crafting recipe now takes 2 diamonds instead of 2 redstone.
 -   You can now select a "Scheduling mode" in the Constructor: first available, round robin, random.
+-   The "applicable upgrades" tooltip on the upgrade slot tooltip now shows the actual items instead of just text.
 
 ### Fixed